src/app/components/products-list/products-list.component.ts
import { Component, OnInit } from '@angular/core';
import { Product } from 'src/app/common/product';
import { ProductService } from 'src/app/services/product.service';

@Component({
  selector: 'app-products-list',
  //templateUrl: './products-list.component.html',
  templateUrl: './products-list-grid.component.html',
  styleUrls: ['./products-list.component.css']
})
export class ProductsListComponent implements OnInit{

  products: Product[] = [];

  constructor(private productService: ProductService){}


  ngOnInit(): void {
    this.listProducts();
  }


  listProducts() {
    this.productService.getProductList().subscribe(
      data => {
        this.products = data;
      }
    )
  }

}
